remove homebrew-fork class Tty
* migrate into our codebase at lib/cask/utils/tty.rb
* merge with our monkeypatch
