+++ lib/cask/utils.rb
@@ -1,6 +1,4 @@
-# coding: utf-8
-# see Homebrew Library/Homebrew/utils.rb
-
+require 'cask/utils/tty'
 require 'yaml'
 require 'open3'
 require 'stringio'
@@ -29,13 +27,6 @@ end
 module MacOS
   def release
     version
-  end
-end
-
-# monkeypatch Tty
-class Tty
-  class << self
-    def magenta; color 35; end
   end
 end
